Fix color of neomake signs in sign column
Prior to this there wasn't enough contrast in the sign column on build
errors.

I guess the "real" solution is to submit a PR to the iceberg theme that
links the neomake sign names to the correct colours, but this is easier
for now.

The new plugins are helpful for generating a test pallette of all
colours in a theme, and inspecting the name of the swatch used on the
token under the cursor.